requirements.txt
# WebSocket Python Project Dependencies
websockets>=12.0
massive
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
pytz
yfinance
requests
schedule